[Refactor] bluemage-data.cppを定義して散らばっていた処理を凝集させた
