+++ src/blue-magic/blue-magic-checker.cpp
@@ -35,7 +35,7 @@ void learn_spell(PlayerType *player_ptr, MonsterAbilityType monspell)
         return;
     }
 
-    auto bluemage_data = PlayerClass(player_ptr).get_specific_data<bluemage_data_type>();
+    auto bluemage_data = PlayerClass(player_ptr).get_specific_data<BluemageData>();
     if (!bluemage_data || bluemage_data->learnt_blue_magics.has(monspell)) {
         return;
     }
@@ -53,7 +53,7 @@ void learn_spell(PlayerType *player_ptr, MonsterAbilityType monspell)
     const auto &monster_power = monster_powers.at(monspell);
     if (randint1(player_ptr->lev + 70) > monster_power.level + 40) {
         bluemage_data->learnt_blue_magics.set(monspell);
-        msg_format(_("%sを学習した！", "You have learned %s!"), monster_power.name);
+        msg_print(_("{}を学習した！", "You have learned {}!"), monster_power.name);
         gain_exp(player_ptr, monster_power.level * monster_power.smana);
         sound(SoundKind::STUDY);
         bluemage_data->new_magic_learned = true;
