+++ src/load/player-class-specific-data-loader.cpp
@@ -79,7 +79,7 @@ void PlayerClassSpecificDataLoader::operator()(std::shared_ptr<force_trainer_dat
     }
 }
 
-void PlayerClassSpecificDataLoader::operator()(std::shared_ptr<bluemage_data_type> &bluemage_data) const
+void PlayerClassSpecificDataLoader::operator()(std::shared_ptr<BluemageData> &bluemage_data) const
 {
     if (loading_savefile_version_is_older_than(9)) {
         auto [magic_num1, magic_num2] = load_old_savfile_magic_num();
