+++ src/load/player-class-specific-data-loader.h
@@ -6,7 +6,7 @@
 struct no_class_specific_data;
 struct smith_data_type;
 struct force_trainer_data_type;
-struct bluemage_data_type;
+class BluemageData;
 class MagicEaterDataList;
 struct bard_data_type;
 struct mane_data_type;
@@ -22,7 +22,7 @@ public:
     void operator()(std::shared_ptr<spell_hex_data_type> &spell_hex_data) const;
     void operator()(std::shared_ptr<smith_data_type> &smith_data) const;
     void operator()(std::shared_ptr<force_trainer_data_type> &force_trainer_data) const;
-    void operator()(std::shared_ptr<bluemage_data_type> &bluemage_data) const;
+    void operator()(std::shared_ptr<BluemageData> &bluemage_data) const;
     void operator()(std::shared_ptr<MagicEaterDataList> &magic_eater_data) const;
     void operator()(std::shared_ptr<bard_data_type> &bird_data) const;
     void operator()(std::shared_ptr<mane_data_type> &mane_data) const;
