+++ src/player-info/class-specific-data.h
@@ -7,7 +7,7 @@ struct no_class_specific_data {
 };
 struct smith_data_type;
 struct force_trainer_data_type;
-struct bluemage_data_type;
+class BluemageData;
 class MagicEaterDataList;
 struct bard_data_type;
 struct mane_data_type;
@@ -22,7 +22,7 @@ using ClassSpecificData = std::variant<
     no_class_specific_data,
     std::shared_ptr<smith_data_type>,
     std::shared_ptr<force_trainer_data_type>,
-    std::shared_ptr<bluemage_data_type>,
+    std::shared_ptr<BluemageData>,
     std::shared_ptr<MagicEaterDataList>,
     std::shared_ptr<bard_data_type>,
     std::shared_ptr<mane_data_type>,
