+++ src/status/bad-status-setter.cpp
@@ -127,7 +127,7 @@ bool BadStatusSetter::set_confusion(const TIME_EFFECT tmp_v)
             msg_print(_("あなたは混乱した！", "You are confused!"));
             if (this->player_ptr->action == ACTION_LEARN) {
                 msg_print(_("学習が続けられない！", "You cannot continue learning!"));
-                auto bluemage_data = PlayerClass(player_ptr).get_specific_data<bluemage_data_type>();
+                auto bluemage_data = PlayerClass(player_ptr).get_specific_data<BluemageData>();
                 bluemage_data->new_magic_learned = false;
                 rfu.set_flag(MainWindowRedrawingFlag::ACTION);
                 this->player_ptr->action = ACTION_NONE;
