+++ src/window/main-window-stat-poster.cpp
@@ -180,7 +180,7 @@ void print_state(PlayerType *player_ptr)
 
     case ACTION_LEARN: {
         text = _("学習", "lear");
-        auto bluemage_data = PlayerClass(player_ptr).get_specific_data<bluemage_data_type>();
+        auto bluemage_data = PlayerClass(player_ptr).get_specific_data<BluemageData>();
         if (bluemage_data->new_magic_learned) {
             attr = TERM_L_RED;
         }
